Early working ronin implementation
This commit constitutes the earliest working draft of the ronin utility
which can be used to synchronize files. Ronin only supports a single
strategy right now (rsync) but it could ideally be extended to support
different strategies in the future.

The setup.py script still needs to be tests and still needs some work,
most of its internals were based off of those for supervisor's setup
script and, therefore, there may be some references to supervisor that
still need to be cleaned up.