Improve ronin performance
Ronin performance was slow when using watchdog to watch and poll a
directory for changes when that directory (or a subdirectory thereof)
had a large number of files. Watchdog doesn't support directory
snapshots that discriminate or ignore paths by default so, for now
ronin has a custom observer, emitter, and directory snapshot classes.
By doing this ronin's polling performance has increased for recursive
polling when a subdirectory contains a large number of files (e.g.
100,000+ files) when that subdirectory is excluded.